Improved workaround for shutting down TFTP server without null pointer error:
Calling serveTFTP or listenAndServeTFTP with a canceled context can cause a nil
pointer error. This happends because of a race condition with when *tftp.Server.Serve
sets *tftp.Server.conn and when ts.Shutdown is called.

One option to "fix" this issue is to PR the following into github.com/pin/tftp:
func (s *Server) Shutdown() {
        if !s.singlePort {
                if s.conn != nil {
                        s.conn.Close()
                }
        }
        q := make(chan struct{})
        s.quit <- q
        <-q
        s.wg.Wait()
}

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>