include the POWEROFF_CMD in ipxe builds
This PR allows devices to poweroff during the IPXE stage.  

In Equinix Metal settings, this allows for scheduled customizations, say to userdata, based on device details that can not be known until the on-demand device is selected and provisioned.

In Tinkerbell settings, this allows for IPXE scripts to perform tasks and then conserve power until awoken by the BMC or user.

Adding this capability should be harmless since the capability is already present at later (and depending on the environment, earlier) stages.
Devices have the capability to poweroff at anytime from an IPXE chained boot. Devices can even reboot from within IPXE with the options provided today.