Remove use of docker and nix in build_ipxe.sh:
Instead of having the build_ipxe.sh handle
running in docker or nix, we brake that out
so that it is up to the user of build_ipxe.sh
to get into a nix-shell. The script no longer
calls docker or nix in any way. The purpose of
this is to simplify the build process, adding
features, maintainability, etc.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>