Add dhcp retries to embedded iPXE script:
This resolves issues when an interface
might need some time before the upstream
port is forwarding properly or if the firmware
is slow, etc. I found this to make a Raspberry PI 4b
boot consistently. This was also tested and resolved
consistency issues with Cisco ACI switches using
LACP configured interfaces.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>