Reduce number of rx buffers in bnxt driver source.
This resolve an issue where downloading large files will hang because
the driver stops reading new packets. This issue was caused by commit
`059c4dc` which changed how drivers are loaded by iPXE. A workaround
wherein the number of recieve buffers is reduced from 8 to 2 was
identified in the issue listed in the function comment.

- Added function to build script to set `NUM_RX_BUFFERS` 2 in the
  Broadcom NetXtreme driver source before compilation.